requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
playwright==1.52.0
beautifulsoup4==4.13.4
python-dateutil==2.9.0.post0
pytz==2025.2
